Add missing-tags script
Replace the super complicated shell, python, and github action yaml mess
with simple python script that is easy to run and debug locally.

The script uses the gh tool to get the release tags, filtering out
pre-release tags.

Since we changed the release artifacts names and the install process we
can no longer mirror any version before v0.7.0.

Example usage with current code:

    % ./missing-tags
    []

When v0.7.0 will be available, the script will print:

    % ./missing-tags
    ["v0.7.0"]

The scripts output is JSON to make it easy to consume in github actions.