workflows: Make it work on a fork
When running in a fork, we want to test the fork install script and
releases. This allows contributors to test changes before posting a pull
request.

For example in my fork the install-test workflow will run:

    curl -fsSL https://github.com/nirs/minikube-vmnet-helper/releases/latest/download/install.sh

And the mirror workflow will run:

    ./missing-tags nirs/minikube-vment-helper

Example mirror upstream releases run:
https://github.com/nirs/minikube-vmnet-helper/actions/runs/17839429668

Example install test run:
https://github.com/nirs/minikube-vmnet-helper/actions/runs/17839537903/job/50725161590

With both changes I can be sure that the change works on the upstream
repo by running the tests in my fork.